inline remaining sequence packing functions and lint
